+++ HISTORY.rst
@@ -42,6 +42,9 @@
   metadata type to document type relationship update.
 - Add link to metadata type events.
 - Add support for subscribing to metadata type events.
+- Add link to view the events of a tag.
+- Add support for subscribing to the events of a tag.
+- Add the tag events view permissions to the tag model ACL.
 
 
 3.0.1 (2018-07-08)